refactor: move change_players_mode to player module
